# Reference segmentation functionality