--- a/build.gradle
+++ b/build.gradle
@@ -4,7 +4,7 @@ plugins {
     id 'jacoco'
 }
 
-version = '1.5.2'
+version = '1.5.3'
 
 
 repositories {