--- a/src/main/java/com/astrazeneca/vardict/Main.java
+++ b/src/main/java/com/astrazeneca/vardict/Main.java
@@ -143,6 +143,9 @@ public class Main {
             conf.chimeric = true;
         }
 
+        if (cmd.hasOption("U")) {
+            conf.enableSV = true;
+        }
         if (cmd.hasOption("UN")) {
             conf.uniqueModeSecondInPairEnabled = true;
         }