--- a/src/main/java/com/astrazeneca/vardict/Main.java
+++ b/src/main/java/com/astrazeneca/vardict/Main.java
@@ -144,7 +144,7 @@ public class Main {
         }
 
         if (cmd.hasOption("U")) {
-            conf.enableSV = true;
+            conf.disableSV = true;
         }
         if (cmd.hasOption("UN")) {
             conf.uniqueModeSecondInPairEnabled = true;