--- a/src/main/java/com/astrazeneca/vardict/ReferenceResource.java
+++ b/src/main/java/com/astrazeneca/vardict/ReferenceResource.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 import java.util.*;
 
 public class ReferenceResource {
-
+    private static final int DEFAULT_REFERENCE_EXTENSION = 1200;
     private static ThreadLocal<Map<String, IndexedFastaSequenceFile>> threadLocalFastaFiles = ThreadLocal.withInitial(HashMap::new);
 
     synchronized private static IndexedFastaSequenceFile fetchFasta(String file) {
@@ -26,14 +26,14 @@ public class ReferenceResource {
     }
 
     /**
-     *
+     * Method convert fasta data to array contains header and nucleotide bases.
      * @param fasta path to fasta file
      * @param chr chromosome name of region
      * @param start start position of region
      * @param end end position of region
      * @return array of nucleotide bases in the region of fasta
      */
-    public static String[] retriveSubSeq(String fasta, String chr, int start, int end) {
+    public static String[] retrieveSubSeq(String fasta, String chr, int start, int end) {
         IndexedFastaSequenceFile idx = fetchFasta(fasta);
         ReferenceSequence seq = idx.getSubsequenceAt(chr, start, end);
         byte[] bases = seq.getBases();
@@ -41,22 +41,25 @@ public class ReferenceResource {
     }
 
     /**
-     * Get part of reference sequence
+     * Get part of reference sequence in
      * @param region region of interest
-     * @return reference sequence map: key - position, value - base
+     * @param chrs map of chromosome lengths
+     * @param conf Vardict Configuration
+     * @return reference object contains sequence map: key - position, value - base and seed map
      */
     public static Reference getREF(Region region, Map<String, Integer> chrs, Configuration conf) {
         Reference ref = new Reference();
-        int extension = 15000;
-        return getREF(region, chrs, conf, extension, ref);
+        return getREF(region, chrs, conf, DEFAULT_REFERENCE_EXTENSION, ref);
     }
 
     /**
      * Get part of reference sequence
      * @param region region of interest
-     * @param ref reference
+     * @param chrs map of chromosome lengths
+     * @param conf Vardict Configuration contains number of nucleotides to extend and seed1, seed2 lengths
      * @param extension extension for reference
-     * @return reference sequence map: key - position, value - base
+     * @param ref reference
+     * @return reference object contains sequence map: key - position, value - base and seed map
      */
     public static Reference getREF(Region region, Map<String, Integer> chrs, Configuration conf, int extension, Reference ref) {
         int sequenceStart = region.start - conf.numberNucleotideToExtend - extension < 1 ? 1
@@ -68,14 +71,14 @@ public class ReferenceResource {
             System.err.println("TIME: Getting REF: " + LocalDateTime.now());
         }
 
-        String[] subSeq = retriveSubSeq(conf.fasta, region.chr, sequenceStart, sequenceEnd);
-//        String header = subSeq[0];
+        String[] subSeq = retrieveSubSeq(conf.fasta, region.chr, sequenceStart, sequenceEnd);
+        //Header doesn't used
+        //String header = subSeq[0];
         String exon = subSeq[1];
 
         if (isLoaded(region.chr, sequenceStart, sequenceEnd, ref)) {
             return ref;
         }
-
         Reference.LoadedRegion loadedRegion = new Reference.LoadedRegion(region.chr, sequenceStart, sequenceEnd);
         ref.loadedRegions.add(loadedRegion);
 
@@ -86,6 +89,7 @@ public class ReferenceResource {
             }
             ref.referenceSequences.put(i + sequenceStart, Character.toUpperCase(exon.charAt(i)));
 
+            // Fill the seed map by sequences of seed1 and seed2 length
             if (exon.length() - i > conf.seed1) {
                 String keySequence = exon.substring(i, i + conf.seed1).toUpperCase();
                 ref = addPositionsToSeedSequence(ref, sequenceStart, i, keySequence);
@@ -103,6 +107,14 @@ public class ReferenceResource {
         return ref;
     }
 
+    /**
+     * Method adds the key sequence for current position in the seed map
+     * @param ref Reference object
+     * @param sequenceStart position of sequence in chromosome
+     * @param i current position in exon to add in seed Map
+     * @param keySequence sequence length of seed1 parameter from chromosome
+     * @return updated Reference
+     */
     public static Reference addPositionsToSeedSequence(Reference ref, int sequenceStart, int i, String keySequence) {
         List<Integer> seedPositions = ref.seed.getOrDefault(keySequence, new ArrayList<>());
         seedPositions.add(i + sequenceStart);