… Readme.md
… src/main/java/com/astrazeneca/vardict/Tuple.java
… src/main/java/com/astrazeneca/vardict/Utils.java
… src/main/java/com/astrazeneca/vardict/VarDict.java
… src/test/java/com/astrazeneca/vardict/integrationtests/IntegrationTest.java
… testdata/fastas/hg19.fa.csv
… testdata/fastas/hs38DH.fa.csv
… testdata/integrationtestcases/Simple;hg19.fa;Colo829-18_S3-sort.bam;chr1;9922-10122;--chimeric -p;.txt
… testdata/integrationtestcases/Simple;hg19.fa;Colo829-18_S3-sort.bam;chr1;9922-10122;-p;.txt
… testdata/integrationtestcases/Somatic;hg19.fa;Colo829-18_S3-sort.bam-Colo829-19_S4-sort.bam;chr1;755917-756517;.txt