### Description
This PR adds reference segmentation for support of chimeric reads and SV finding.
**Please, merge this after java 10 fixes (#116).**
### Implementation
- Added `seed1` and `seed2` variables for lengths of parts of reference.
- Added `Reference` class that contains information about loaded regions, reference for this part of fasta and seed parts of seed1 and seed2 length. Objects of this class are used in `toVars()`, `parseSam()` and other methods to get seed parts and already loaded regions.
- Default reference extension is set to 1200 bp.
- Changes in `findconseq()`, `isMatch()` and `isMatchRef()` methods as they implemented in Perl (except adaptor finding).
- Removed variable `buffer`: now it is not used while processing sclip 3' and 5' (as in Perl).
- Test cases references lengths were changed due to bigger extension (no additional test cases because the reference is checked implicitly in all test cases).

### Performance
Na12878 test used bam and bed: ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/phase3/data/NA12878/exome_alignment/NA12878.mapped.ILLUMINA.bwa.CEU.exome.20121211.bam
ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/technical/reference/exome_pull_down_targets_phases1_and_2/20120518.exome.consensus.bed

Bam | Master, Elapsed (wall clock) time | reference_segmentation branch, Elapsed (wall clock) time
-- | -- | --
NA12878 16,1Gb | 7 m 58 s | 8 m 3 s

The difference in performance is within the margin of error.